+++ setup.py
@@ -85,6 +85,7 @@ if __name__ == '__main__':
             'Programming Language :: Python :: 3.4',
             'Programming Language :: Python :: 3.6',
             'Programming Language :: Python :: 3.7',
+            'Programming Language :: Python :: 3.8',
             'License :: OSI Approved :: Apache Software License',
             'Topic :: Software Development :: Testing',
         ]
